+++ graphein/protein/visualisation.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 import re
 from itertools import count
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Callable, Dict, List, Optional, Tuple, Union
 
 import matplotlib
 import matplotlib.pyplot as plt
@@ -22,6 +22,7 @@ import seaborn as sns
 from loguru import logger as log
 from mpl_toolkits.mplot3d import Axes3D
 
+from graphein.protein.resi_atoms import HYDROPHOBICITY_SCALES
 from graphein.protein.subgraphs import extract_k_hop_subgraph
 from graphein.utils.dependencies import import_message
 
@@ -45,6 +46,122 @@ except ImportError:
         extras=True,
     )
     log.debug(message)
+
+
+"""
+TODO: Functino that gets ``min`` and ``max`` values in a graph for a given feature so that we can scale / offset to > 0
+TODO: should feature `distance` actually contain the site itself i.e. in the string?
+"""
+
+
+def _node_feature_func(
+    g: nx.Graph,
+    feature: str,
+    focal_node: Optional[str] = None,
+    focal_point: Optional[tuple] = None,
+    no_negatives: bool = False,
+) -> Callable:
+    """
+    Maps a feature as described by a string to a function that can be applied on nodes from a graph.
+
+    :param g: Protein graph.
+    :type g: nx.Graph
+    :param feature: Name of feature to extract.
+    :type feature: str
+    :param focal_node: A specific node within ``g`` to use in feature calculation; e.g. when calculating ``distance`` to a given site.
+    :type focal_node: Optional[str]
+    :param focal_point: Use specific coordinates instead of a node within the graph.
+    :type focal_point: tuple
+    :param no_negatives: Take the max of ``0`` and the feature's value. Defaults to ``False``.
+    :type no_negatives: bool
+    :return: Function that returns a value for a given node ID.
+    :rtype: Callable
+
+    TODO is there a way to wrap a lambda with another function i.e. max(0, f) for `no_negatives` ?
+    """
+    if feature == "degree":
+        return lambda k: g.degree[k]
+    elif feature in ["seq-position", "seq_position"]:
+        return lambda k: g.nodes(data=True)[k]["residue_number"]
+    elif feature == "rsa":
+        return lambda k: g.nodes(data=True)[k]["rsa"]
+    elif feature in ["bfac", "bfactor", "b_factor", "b-factor"]:
+        return lambda k: g.nodes(data=True)[k]["b_factor"]
+    elif (
+        feature == "distance"
+    ):  # Euclidean distance to a specific node / coordinate
+
+        def get_coords(g: nx.Graph, node: str) -> np.ndarray:
+            return np.array(g.nodes()[node]["coords"])
+
+        if focal_node:
+            assert focal_node in g.nodes()
+            return lambda k: np.linalg.norm(
+                get_coords(g, k) - get_coords(g, focal_node)
+            )
+        elif focal_point:
+            assert len(focal_point) == 3
+            return lambda k: np.linalg.norm(
+                get_coords(g, k) - np.array(focal_point)
+            )
+        else:
+            raise ValueError(
+                f"Node feature 'distance' requires one of `focal_node` or `focal_point`."
+            )
+
+    # Meiler embedding dimension
+    p = re.compile("meiler-?([0-9])")
+    match = p.search(feature)
+    if match:
+        dim = match.group(1)
+        if int(dim) in range(1, 8):
+            if no_negatives:
+                return lambda k: max(
+                    0, g.nodes(data=True)[k]["meiler"][f"dim_{dim}"]
+                )
+            else:
+                return lambda k: g.nodes(data=True)[k]["meiler"][f"dim_{dim}"]
+        else:
+            raise ValueError(
+                f"Meiler embeddings have dimensions 1-7, received {dim}."
+            )
+
+    # Hydrophobicity
+    p = re.compile(
+        "([a-z]{2})?-?(hydrophobicity)"
+    )  # e.g.  "kd-hydrophobicity", "tthydrophobicity", "hydrophobicity"
+    match = p.search(feature)
+    if match and match.group(2):
+        # TODO: check if nodes actually have 'hydrophobicity' already; if they do, then use this.  if not, then map to kd.
+        scale: str = (
+            match.group(1) if match.group(1) else "kd"
+        )  # use 'kdhydrophobicity' as default if no scale specified
+        try:
+            hydrophob: Dict[str, float] = HYDROPHOBICITY_SCALES[scale]
+        except:
+            raise KeyError(f"'{scale}' not a valid hydrophobicity scale.")
+        return lambda k: hydrophob[k.split(":")[1]]
+    else:
+        raise NotImplementedError(f"Feature '{feature}' not implemented.")
+
+
+def _node_size_func(
+    g: nx.Graph, feature: str, min: float, multiplier: float
+) -> Callable:
+    """
+    Returns a function that can be use to generate node sizes for plotting.
+
+    :param g: Protein graph
+    :type g: nx.Graph
+    :param feature: Name of feature to scale node sizes by.
+    :type feature: str
+    :param min: Number to offset size with.
+    :type min: float
+    :param multiplier: Number to scale feature values by.
+    :type multiplier: float
+    """
+    get_feature = _node_feature_func(g=g, feature=feature, no_negatives=True)
+    return lambda k: min + multiplier * get_feature(k)
 
 
 def plot_pointcloud(mesh: Meshes, title: str = "") -> Axes3D:
@@ -255,27 +372,12 @@ def plotly_protein_structure_graph(
         G, colour_map=edge_color_map, colour_by=colour_edges_by
     )
 
-    # Get node size
-    def node_scale_by(G: nx.Graph, feature: str):
-        if feature == "degree":
-            return lambda k: node_size_min + node_size_multiplier * G.degree[k]
-        elif feature == "rsa":
-            return (
-                lambda k: node_size_min
-                + node_size_multiplier * G.nodes(data=True)[k]["rsa"]
-            )
-
-        # Meiler embedding dimension
-        p = re.compile("meiler-([1-7])")
-        dim = p.search(feature)[1]
-        if dim:
-            return lambda k: node_size_min + node_size_multiplier * max(
-                0, G.nodes(data=True)[k]["meiler"][f"dim_{dim}"]
-            )  # Meiler values may be negative
-        else:
-            raise ValueError(f"Cannot size nodes by feature '{feature}'")
-
-    get_node_size = node_scale_by(G, node_size_feature)
+    size_by = _node_size_func(
+        G,
+        node_size_feature,
+        min=node_size_min,
+        multiplier=node_size_multiplier,
+    )
 
     # 3D network plot
     x_nodes = []
@@ -289,7 +391,7 @@ def plotly_protein_structure_graph(
         x_nodes.append(value[0])
         y_nodes.append(value[1])
         z_nodes.append(value[2])
-        node_sizes.append(get_node_size(key))
+        node_sizes.append(size_by(key))
 
         if label_node_ids:
             node_labels.append(list(G.nodes())[i])
@@ -736,6 +838,7 @@ def asteroid_plot(
     node_id: str,
     k: int = 2,
     colour_nodes_by: str = "shell",  # residue_name
+    size_nodes_by: str = "degree",
     colour_edges_by: str = "kind",
     edge_colour_map: plt.cm.Colormap = plt.cm.plasma,
     edge_alpha: float = 1.0,
@@ -746,6 +849,8 @@ def asteroid_plot(
     use_plotly: bool = True,
     show_edges: bool = False,
     show_legend: bool = True,
+    node_size_min: float = 20,
+    node_size_multiplier: float = 10,
     node_size_multiplier: float = 10.0,
 ) -> Union[plotly.graph_objects.Figure, matplotlib.figure.Figure]:
     # sourcery skip: remove-unnecessary-else, swap-if-else-branches
@@ -763,6 +868,8 @@ def asteroid_plot(
     :param colour_nodes_by: Colour the nodes by this attribute. Currently only
         ``"shell"`` is supported.
     :type colour_nodes_by: str
+    :param size_nodes_by: Size the nodes by an attribute.
+    :type size_nodes_by: str
     :param colour_edges_by: Colour the edges by this attribute. Currently only
         ``"kind"`` is supported.
     :type colour_edges_by: str
@@ -784,6 +891,8 @@ def asteroid_plot(
     :param show_legend: Whether to show the legend of the edges. Defaults to
         `True``.
     :type show_legend: bool
+    :param node_size_min: Specifies node minimum size. Defaults to ``20.0``.
+    :type node_size_min: float
     :param node_size_multiplier: Multiplier for the size of the nodes. Defaults
         to ``10.0``.
     :type node_size_multiplier: float
@@ -845,21 +954,34 @@ def asteroid_plot(
             node_x.append(x)
             node_y.append(y)
 
-        degrees = [
-            subgraph.degree(n) * node_size_multiplier for n in subgraph.nodes()
-        ]
+        size_by = _node_size_func(
+            subgraph,
+            size_nodes_by,
+            min=node_size_min,
+            multiplier=node_size_multiplier,
+        )
+        node_sizes = [size_by(n) for n in subgraph.nodes()]
 
+        colour_nodes_by = colour_nodes_by.lower()
+        node_colours = []
         if colour_nodes_by == "shell":
-            node_colours = []
             for n in subgraph.nodes():
                 for k, v in nodes.items():
                     if n in v:
                         node_colours.append(k)
         else:
-            raise NotImplementedError(
-                f"Colour by {colour_nodes_by} not implemented."
-            )
-            # TODO colour by AA type
+            try:
+                get_feature = _node_feature_func(
+                    g=subgraph, feature=colour_nodes_by, no_negatives=False
+                )
+            except:
+                raise NotImplementedError(
+                    f"Colour by {colour_nodes_by} not implemented."
+                )
+
+            for n, d in subgraph.nodes(data=True):
+                node_colours.append(get_feature(n))
+
         node_trace = go.Scatter(
             x=node_x,
             y=node_y,
@@ -869,13 +991,13 @@ def asteroid_plot(
             textposition="bottom center",
             showlegend=False,
             marker=dict(
-                colorscale="YlGnBu",
+                colorscale="viridis",
                 reversescale=True,
                 color=node_colours,
-                size=degrees,
+                size=node_sizes,
                 colorbar=dict(
                     thickness=15,
-                    title="Shell",
+                    title=str.capitalize(colour_nodes_by),
                     tickvals=list(range(k)),
                     xanchor="left",
                 ),